removed redundant '-X POST' from curl example
When -F is used with curl it will do a POST so using '-X POST' is
redundant in this case. If used together with -L would possible
break what we like to achive.

Here's a video where Daniel Stenberg, project leader and main developer
for the cURL project explains why we shouldn't use '-X POST' excessively
https://youtu.be/y7Mxq013Dy4?t=690