update systemd (file -> append)
The value `file` for `StandardOutput` and `StandardError` works differently than expected.
The file is created and used only once - ever. This means that when the service is restarted,
there won't be any logging anymore.
Using the keyword `append` will reuse the log file and is the correct way to specify logfiles in systemd.